fix: move new role system endpoints to oss
