+++ packages/backend/src/ee/controllers/CustomRolesController.ts
@@ -1,9 +1,12 @@
 import {
     AddScopesToRole,
+    ApiDefaultRoleResponse,
     ApiErrorPayload,
     ApiRemoveScopeFromRoleResponse,
     ApiRoleWithScopesResponse,
     ApiUnassignRoleFromUserResponse,
+    CreateRole,
+    UpdateRole,
 } from '@lightdash/common';
 import {
     Body,
@@ -27,24 +30,24 @@ import {
     unauthorisedInDemo,
 } from '../../controllers/authentication';
 import { BaseController } from '../../controllers/baseController';
-import { RolesService } from '../../services/RolesService/RolesService';
 
 /**
- * This controller handles all role management operations with the new namespace structure:
- * Check ProjectRolesController.ts and OrganizationRolesController.ts for using these roles.
+ * EE Custom Roles API
  *
- * Role Management: /api/v2/roles
- * - Creating, reading, updating, and deleting roles
+ * This API is only available for enterprise customers.
+ * EE custom role management endpoints under organization scope:
+ * - POST / - Create custom organization role
+ * - PATCH /{roleUuid} - Update custom organization role
+ * - DELETE /{roleUuid} - Delete custom organization role
+ * - POST /{roleUuid}/scopes - Add scopes to custom organization role
+ * - DELETE /{roleUuid}/scopes/{scopeName} - Remove scope from custom organization role
  *
- * Role scope management: /api/v2/roles/{roleUuid}/scopes
- * - Adding scopes to roles
- * - Removing scopes from roles
- *
+ * For read-only operations (GET endpoints), see the OSS Organization Roles API
  */
-@Route('/api/v2/roles')
+@Route('/api/v2/orgs/{orgUuid}/roles')
 @Response<ApiErrorPayload>('default', 'Error')
-@Tags('v2', 'Roles')
-export class RolesController extends BaseController {
+@Tags('v2', 'Custom Roles')
+export class CustomRolesController extends BaseController {
     /**
      * Convenience method to access the roles service without having
      * to specify an interface type.
@@ -54,25 +57,86 @@ export class RolesController extends BaseController {
     }
 
     /**
-     * Get role by ID
+     * Create a new role in organization
      */
-    @Middlewares([allowApiKeyAuthentication, isAuthenticated])
-    @SuccessResponse('200', 'Success')
-    @Get('/{roleUuid}')
-    @OperationId('GetRoleById')
-    async getRoleById(
+    @Middlewares([
+        allowApiKeyAuthentication,
+        isAuthenticated,
+        unauthorisedInDemo,
+    ])
+    @SuccessResponse('201', 'Created')
+    @Post()
+    @OperationId('CreateOrganizationRole')
+    async createOrganizationRole(
         @Request() req: express.Request,
+        @Path() orgUuid: string,
+        @Body() body: CreateRole,
+    ): Promise<ApiDefaultRoleResponse> {
+        const role = await this.getRolesService().createRole(
+            req.account!,
+            orgUuid,
+            body,
+        );
+
+        this.setStatus(201);
+        return {
+            status: 'ok',
+            results: role,
+        };
+    }
+
+    /**
+     * Update role in organization
+     */
+    @Middlewares([
+        allowApiKeyAuthentication,
+        isAuthenticated,
+        unauthorisedInDemo,
+    ])
+    @SuccessResponse('200', 'Success')
+    @Patch('/{roleUuid}')
+    @OperationId('UpdateOrganizationRole')
+    async updateOrganizationRole(
+        @Request() req: express.Request,
+        @Path() orgUuid: string,
         @Path() roleUuid: string,
-    ): Promise<ApiRoleWithScopesResponse> {
-        const role = await this.getRolesService().getRoleByUuid(
+        @Body() body: UpdateRole,
+    ): Promise<ApiDefaultRoleResponse> {
+        const role = await this.getRolesService().updateRole(
             req.account!,
             roleUuid,
+            body,
         );
 
         this.setStatus(200);
         return {
             status: 'ok',
             results: role,
+        };
+    }
+
+    /**
+     * Delete role from organization
+     */
+    @Middlewares([
+        allowApiKeyAuthentication,
+        isAuthenticated,
+        unauthorisedInDemo,
+    ])
+    @SuccessResponse('200', 'Success')
+    @Delete('/{roleUuid}')
+    @OperationId('DeleteOrganizationRole')
+    async deleteOrganizationRole(
+        @Request() req: express.Request,
+        @Path() orgUuid: string,
+        @Path() roleUuid: string,
+    ): Promise<ApiUnassignRoleFromUserResponse> {
+        await this.getRolesService().deleteRole(req.account!, roleUuid);
+
+        this.setStatus(200);
+        return {
+            status: 'ok',
+            results: undefined,
         };
     }
 
@@ -89,6 +153,7 @@ export class RolesController extends BaseController {
     @OperationId('AddScopesToRole')
     async addScopesToRole(
         @Request() req: express.Request,
+        @Path() orgUuid: string,
         @Path() roleUuid: string,
         @Body() body: AddScopesToRole,
     ): Promise<ApiUnassignRoleFromUserResponse> {
@@ -118,6 +183,7 @@ export class RolesController extends BaseController {
     @OperationId('RemoveScopeFromRole')
     async removeScopeFromRole(
         @Request() req: express.Request,
+        @Path() orgUuid: string,
         @Path() roleUuid: string,
         @Path() scopeName: string,
     ): Promise<ApiRemoveScopeFromRoleResponse> {
